Moved has_community_services_card into social_security
